Continuing to adapt 'Bash Shell - Command List'
Currently got to the "sed" command, which was a complete mess in the old
notes, but I'm actually figuring out how to clean it and explain it in a
clear way.
The separation between "kinds of arguments" with the arg callout helps.
With 1 arg for options and 1 arg for actions, (which I haven't done
yet), I can focus on 1 walkthrough callout to get to the general
behavior of sed. And by adding examples in this callout, I can keep the
actual sed explanation pretty simple. I should probably include the
numbered list organization in this walkthrough.

Also added a footnote for "pattern space" in the "sed" section, and the
"patch" section now redirects directly to the "diff" section through a
subscript internal link. I couldn't make the word "diff" directly a link
and have it in a code block at the same time, so I decided to proceed
with a subscript link number. (A superscript would have visually
conflicted with footnotes). I guess I'm gonna stick with this way when
refering to a command outside of its own section.

Decided to keep both callouts (walkthrough and example) : If an example
can be summed up in one sentence/line, then it is not an explanation
step by step, then I write it BEFORE the code example. If it is not the
case, then I am just going to nest a walkthrough callout inside the
Example callout.
On the callout theme, I also created a new callout, "goodprac" (usually
renamed "Bonne Pratique" for "Good Practice"). Name should be pretty
self-explanatory.
So right now my custom callouts list is :
	- walkthrough
		Step by step explanation of preceding code or concept.
		Usually organized in a numbered list.
	- example
		Concrete code example, with simple explanation or a
		Walkthrough callout included.
	- arg
		Any kind of arguments the command takes. Usually divided
		in 1 callout per argument kind. (e.g. one for "options",
		one for "actions")
	- goodprac
		Little tips of good practice.

I'm going to try and write detailed commit message from now on, to try
and get used to it, and we never know, might be useful.