Moved/adapted 'Bash Shell - Wildcards and Regex'
Didn't have much to do, was already well organized with bullet point
lists.
Did go back to 'Bash Shell - Command List' to add a subscript link to
the regex section of this note on every command that mentionned the use
of regex. Didn't know if I should put it at EVERY occurence of 'regex',
or if on the contrary I should just put it on the first, so I went in
the middle and put it on the first occurence of 'regex' in EVERY command
that mentionned its use.
Decided subscript internal links are not going to be numbers anymore,
they're going to conflict ith the footnotes which are ALSO numbers. So
now they're just the name of the thing they're referring to, but the
problem is, it's kind of redundant because usually it's already the word
itself that the link follows. Here the mentions of 'regex' are
immediatly followed by a little subscript 'regex'. Hoping I'll find a
better way to write this.