Finished moving and adapting "Git - Overview"
Was pretty straightforward actually. There wasn't any code, it was only
pure text, that I almost didn't have to rearrange, except for the 2
examples in the "branches" section, which I obviously put in Example
callouts. Other than that it was nothing more than cleaning the mess
that moving the note creates.
Now, onto "Git - Comman List", for moving, adapting, and finally
finishing it, as it was the one I was working on when I decided to move
the notes.