FINISHED WRITING 'Git - Command List' !!!
I added fetch, pull, checkout, restore, and merge. I decided to let out
reset and rebase, as I CLEARLY will not need them before a long time, so
I guess I'll assess when that time comes.
The fetch section mad me modify the branches section of 'Git -
Overview', creating a subsection 'Fonctionnement général des branches',
which includes everything that was in the branch section before, and I
added another subsection called 'remote-tracking branch' to explain the
concept of remote-tracking branches (duh). And so I could mention the
remote-tracking branch in the fetch section, and add as always a
subscript internal link to this Overview section.
The pull section went smoothly.
For the checkout section, I initially mentionned the possibility
of restoring files, but then I realised `restore` what precisely made for this
purpose. I decided it was easier for me to have separate commands with specific
uses, so the checkout section doesn't even mention the possibility of
restoring files.
The restore section went smoothly.
The merge section... Well, I'll admit I kind of rushed it. It was either
I went on a mission to understand it at a super deep level and write a
suuuper long section about it, or I decided to not care. As I've been
stuck on the Bash and Git part for too long now and I just want to get
to C, I decided to not care and write a short basic description. Maybe
I'll take care of that later.