Started moving and adapting 'C - Overview'
Okay, so first, it wasn't called 'C - Overview' but 'Notions C', but for
consistency I renamed it on the same pattern than for git and bash.
So, when I finished with git and bash, I could finally start with C, but
I didn't remember that also had notes in C in Evernote. But that's
actually a good thing, having to move and adapt them in obsidian helps
me learn them. And I tried watching the school's videos, but god are
they awfully explained, my notes are much better lol
So anyway, I have done "Variables", "File Descriptor", "Opérateurs",
"Conditions", and "Boucles". I started the last 2 ("Fonctions" and
"Pointeurs"), but it's already late today.

I also noticed there was another note called "Fonctions C", with only
one function described, so I'm guessing I just forgot to complete it,
but on the same pattern as the "Command List" for git and bash, it could
become a "Function List" later.