Finished writing basic 'Function' section
I was adapting 'C - Overview', and I realised the 'functions' section
was pretty bad, and it didn't have the right goal. Since functions are
pretty much everything in C, and everything is structured around them, I
felt it should be the first section, as well as the most thorough
section, so that the goal is that when I get back to C after a long
time, I can still understand everything there is to know about functions
in C just with this section. So I rewrote it completely in a much better
way, and I plan to expand it as I keep learning, and I plan for this
section to be the most important and thorough of my whole notes.
Also added a paragraph about the main() function, as there was no other
section it could be in.