Added stuff to C - Overview/Created Function List
Modified the Function section of 'C - Overview', made it into different
subsection, one for the main() function, one for declaring functions in
general, which were both already there before this, and added the
subsection about prototypes (or how to include and call functions).

I also created the note 'C - Function List' as mentionned in a previous
commit, indeed taking the write function that was already there. I'm not
sure yet as to how I'm gonna organize these explanation, I'm not really
happy with it right now, but we'll see anyway.
Also added a subscript internal link from the write section, at the
"File Descriptor" mention, linking to its section in 'C - Overview'

I also added a 'To do' section at the end, with the topics I haven't
tackled yet, but I will have to, just to be sure not to forget them.
Currently only the keywords 'static' and 'const' are in it, because I
don't need them right now and I don't know where I'd put them, I don't
know if it'd be worth it to just create a whole section for the both of
them, and how to call it, anyway, that's why I just stashed them here
for now.