Added 1 function to the list, and changed sublinks
I changed the organisation of 'Function List', because I realised I need
to put already existing functions from already existing libraries, AS
WELL AS my own functions, those that 42 school makes me create and that
I might use later
So I created to big sections : 'Custom Functions' for my own and
'Existing Libraries', which will itself be divised into 1 section per
library.
So I added in 'Custom' my ft_putchar(), its use, declaration, and list
of function used. I created for this 2 new callouts, 'Declaration', for
the code of the declaration, and 'List', in order to list the functions
used, but the name allows it to maybe be used for something else in the
future. In this use, the 'list' callout is included inside the
'declaration' callout, after the code.

I also changed the syntax of subscripts internal links, where before the
subscript text was usually the exact same word it was following, which
was weird and redundant, so I decided to just use one common word
'more', as just a keyword meaning the link will bring me to more
information about the preceding word.