Simple typo correction in Git - Command List
Although I also realized I did very detailed commits last year with the notes, and I just went through them all, and it proved VERY useful actually, so Imma start doing it again. And to do that, I used the commit title and commit body, which, since then I forgot how to do, so even though this is an extremely minor commit, this is mainly a test. I hope this is the commit body. I mean obviously it is judging from the length of this text, but I hope the separation between the title and the body is clean. This also serves as a statement that from now on, except from minor correction commits like this one, every commit is gonna be detailed about what I did there.