Yet again, changed the subscripts internal links organization
I realized it was completely dumb to add a subscript more to serve as the internal link AFTER the word that needed an internal link, when I can just turn that word itself into the link. It doesn't work with when it's in a code block or even an inline code, (which is the reason I initially search for this kind of solutions), so I'll keep doing that for those, but I had absolutely no reason to keep this syntax consistent when the word is not in a code block. So I changed most of them that way. Might have to check other notes to see if I did that there as well.