Added ft_strcpy() and ft_strncpy() to Function List
Added the prototype, definition (including special cases), and declaration of the functions ft_strcpy() and ft_strncpy() to Function Lists. I also modified a bit the Empty line spacings between paragraphs, but I'm doing another commit about that right after. But the modifications of empty lines concerning this note are also included in this commit.