Changed font size appearance and the 'empty line spacings' organisation
Reduced the font size appearance because it was just too big. Also tried to establish a pattern regarding the empty lines I use to separate paragraphs, and also my use of <br>, which I have no idea what is does, but it appears to be doing some kind of newline/breakline, however the specifics are a complete mystery. But I realised I could stack the <br> on the same 'Source' line, and it indeed stacked new lines in Reading. So i decided to go for basically, depending on the number of levels of subsections, either 2 or 3 <br> between each #paragraph, 1 or 2 between each ##paragraph, etc. I don't know if I'll stick to it or anything but I'll try it at least.