Added 2 "info" Callouts, and detailed a section in Git - Overview
I added an info callout in 'C - Overview', in the 'Operateurs' section,
about the priorities between the '&&' operator and the '||' operator.
I also added one in 'Git - Command List', in the `git commit` section,
about how to separate title and message with the `-m` option. I also
subsequently changed the description of the command, where it specified
that just typing `git commit` would open an editor, and explained the
same, but for the editor.

Those first 2 changes are just small details I encountered while
working, but the other change is more important.
I realized doing this for the second time, I already remembered most of
what I learned last year, and the exceptions were mostly covered by the
notes. But not entirely, and that's where I asked chatGPT a lot of
things, and that's precisely where I would loose a lot of time. And
that's actually kind of what I can do for this 2nd edition (at least I
feel like it would be the most useful thing) :
Going back through my chat GPT conversation and progressively try to
cover everything in the notes. Since, if I had to go to chatGPT, it
means my notes were lacking on that topic. The goal would be for my
notes, that I could do what I've been doing while only relying on my
notes.

So this one is the first and most following commits should be about
this.
Here I detailed the 'remote-tracking branch' subsection in 'Git -
Overview'. More specifically, I added information about how to check
your remote-tacking branches with the `git branch`.

I think my Overview, especially the Git one, are a little bit too
focused on the theory. Like, theory is important and should remain the
main focus of these pages, but it shouldn't be *ONLY* that, or it's going
to be completely disconnected knowledge.