Added a warning for `bc` in 'Command List' and an info for 'if' in 'C - Overview' :
First I added a warning callout for the `bc` command in 'Bash Shell -
Command List' because it did not specify that the order in which you
affect 'ibase' and 'obase' is DEFINITELY important, because the moment
you affect 'ibase', it's effective immediatly, thus reading further
inputs in new base, even if on the same command line.

And secondly, I added an info callout for if and while in 'C - Overview'
about the fact that you can use functions as conditions if they return 0
or 1.
I mean, the info callout is only in the if section, but there's a call
back in the while section.
Also proceeded to correct a few typos while I was at it, and I removed
a completely useless example for if, and a few other small adjustments.