Added an Info in the Pointer section of 'C - Overview'
(Will start calling 'blabla callout' just 'Blabla' instead in commit
messages)

Added an info about how address are just non negative ints, like char
are just ints, but interpreted as a character through the ASCII table.

Also, while doing that I noticed the way I organized my paragraphs was
weird, so I rearranged it for the pointer section, and I'm gonna do
that for the rest of the notes, but I'm going to do a dedicated commit
for this where I describe the way I did things. So just assume whatever
I did to the paragraphs in the pointer section in this commit
corresponds to what I describe in the next commit.

And lastly, this is actually the last commit of "going through my
chatGPT conversation", I thought there was much more stuff than that,
but it turns out most of the questions I asked where very exercise
specific and didn't give much knowledge about coding in a broader
understanding.