Cleaned syntax/paragraph for 'C - Overview'
I don't know if I'm going to do it for every note, but that was the most
important one.
Mainly it was because there were paragraphs that were much too long and
thus unreadable, and also I went too far with my bullet points, so I
dialed it back a bit so it looks cleaner overall.
Other than that, small corrections. 2 bigger ones :
- I reorganised the file type list, and added a mention of the pointers
  (since it's technically a variable), which... Points to the
  pointer section...
- In 'Operateurs' section, I rewrote completely the part where I mention
  post-incrementation / decrementation, because it didn't even mention
  the post/pre stuff, it literally just said "oh you can increment even
  easier".