Corrected an important mistake in C Overview, Pointer section
Corrected a straight up false statement in the Pointer section of C - Overview. While looking for another info, I realized I wrote that to declare a pointer, you had to declare it as a `int *ptr` because 'since it stores an adress, which is a number, a pointer is then technically an int', which is just horrendously wrong. A pointer is a variable type. Meaning a pointer is a variable of type... Well pointer. Every fucking variable stores a number value, the type is here to tell the compiler HOW to interpret this variable. So the type you put before the * when declaring a pointer is just the type of the variable POINTED by the pointer. If you want to declare a pointer to a char variable, then you declare a `char *ptr`. I don't understand how that got in my notes, as it looks like just a random guess as to how pointers work.