Added ft_atoi() to Function List, made an adjustment to the overall syntax, and a small correction to ft_putnbr()
I realised pretty late that ft_atoi() is definitely one of the functions
to add to my 'C - Function List' (late as in I made it like 1 or 2 weeks
ago), because now that I use the main arguments argc and argv to do my
testing, if I need to use ints, argv is always a string so I need atoi
to convert it into an int. So I added it, and instead of writing the
specification of how it works and what to keep wary about, I realised
they were very detailed in the instructions for the exercise so I just
copied them and modified them a little and it covered everything
necessary so that was it.
Also while writing the general description of the function, and when
looking at how I wrote the others, I realised my syntax was completely
dumb and too "over the top". Like, every time the function took a string
as an argument, for example `char *str`, in the description I designated
it by "the content of the string pointed by str", instead of just going
"the string str". I don't know why I did that, but now it's fixed in the
whole page.
Also my second declaration of ft_putnbr() was missing a ";" at one
point, so corrected that.